# backend/docs/swagger.json
{
  "schemes": ["http"],
  "swagger": "2.0",
  "info": {
    "description": "AI Agent Demo — Go + Gin 后端，提供 13 个技术演示场景的 API 支持",
    "title": "Interview Demo API",
    "contact": {
      "name": "Developer",
      "email": "dev@example.com"
    },
    "version": "1.0"
  },
  "host": "localhost:8080",
  "basePath": "/api",
  "paths": {
    "/agents": {
      "get": {
        "description": "返回所有已创建的智能体列表",
        "produces": ["application/json"],
        "tags": ["智能体"],
        "summary": "列出智能体",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      },
      "post": {
        "description": "创建指定类型的 AI 智能体（ReAct/Function/Multi/RAG）",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["智能体"],
        "summary": "创建智能体",
        "parameters": [
          {
            "description": "创建请求",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "type": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "/agents/{id}": {
      "delete": {
        "description": "删除指定的智能体",
        "produces": ["application/json"],
        "tags": ["智能体"],
        "summary": "删除智能体",
        "parameters": [
          {
            "type": "string",
            "description": "智能体 ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "/agents/{id}/execute": {
      "post": {
        "description": "执行指定智能体，返回推理过程和最终响应",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["智能体"],
        "summary": "执行智能体",
        "parameters": [
          {
            "type": "string",
            "description": "智能体 ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "执行请求",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "input": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
...
    "/auth/check": {
      "get": {
        "description": "验证 JWT Token 是否有效，返回剩余过期时间",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["认证"],
        "summary": "验证 Token",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "description": "使用用户名密码登录，获取 JWT Token",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["认证"],
        "summary": "用户登录",
        "parameters": [
          {
            "description": "登录请求",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/auth.LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/auth.TokenResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "description": "使用 Refresh Token 轮换获取新的 Access Token 和 Refresh Token，带重放攻击检测",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["认证"],
        "summary": "刷新 Token",
        "parameters": [
          {
            "description": "刷新请求",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/auth.RefreshRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
...
    "/chat": {
      "post": {
        "description": "发送消息并获取 AI 回复，支持知识库上下文、Agent 调用、模型选择",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["对话"],
        "summary": "发送聊天消息",
        "parameters": [
          {
            "description": "聊天请求",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "agentId": {
                  "type": "string"
                },
                "agentType": {
                  "type": "string"
                },
                "content": {
                  "type": "string"
                },
                "conversationId": {
                  "type": "string"
                },
                "knowledgeBaseId": {
                  "type": "string"
                },
                "model": {
                  "type": "string"
                },
                "useAgent": {
                  "type": "boolean"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "/chat/history/{conversationId}": {
      "get": {
        "description": "获取指定会话的聊天历史消息",
        "produces": ["application/json"],
        "tags": ["对话"],
        "summary": "获取聊天历史",
        "parameters": [
          {
            "type": "string",
            "description": "会话 ID",
            "name": "conversationId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      },
      "delete": {
        "description": "清空指定会话的聊天历史消息",
        "produces": ["application/json"],
        "tags": ["对话"],
        "summary": "清空聊天历史",
        "parameters": [
          {
            "type": "string",
            "description": "会话 ID",
            "name": "conversationId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "/chat/stream": {
      "post": {
        "description": "通过 Server-Sent Events 返回流式聊天回复",
        "consumes": ["application/json"],
        "produces": ["text/event-stream"],
        "tags": ["对话"],
        "summary": "聊天流式回复 (SSE)",
        "parameters": [
          {
            "description": "聊天请求",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "content": {
                  "type": "string"
                },
                "conversationId": {
                  "type": "string"
                },
                "knowledgeBaseId": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
...
    "/models/{id}": {
      "get": {
        "description": "获取指定 AI 模型的详细信息",
        "produces": ["application/json"],
        "tags": ["模型"],
        "summary": "获取模型详情",
        "parameters": [
          {
            "type": "string",
            "description": "模型 ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "/models/{id}/chat": {
      "post": {
        "description": "使用指定模型进行对话",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["模型"],
        "summary": "模型对话",
        "parameters": [
          {
            "type": "string",
            "description": "模型 ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "对话请求",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "messages": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/model.Message"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "auth.LoginRequest": {
      "type": "object",
      "required": ["password", "username"],
      "properties": {
        "password": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      }
    },
    "auth.RefreshRequest": {
      "type": "object",
      "required": ["refresh_token"],
      "properties": {
        "refresh_token": {
          "type": "string"
        }
      }
    },
    "auth.TokenResponse": {
      "type": "object",
      "properties": {
        "access_token": {
          "type": "string"
        },
        "expires_in": {
          "type": "integer"
        },
        "refresh_token": {
          "type": "string"
        }
      }
    },
...
    "model.Message": {
      "type": "object",
      "required": ["content", "role"],
      "properties": {
        "content": {
          "type": "string"
        },
        "conversation_id": {
          "type": "string"
        },
        "created_at": {
          "type": "string"
        },
        "id": {
          "type": "string"
        },
        "role": {
          "type": "string"
        },
        "timestamp": {
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "description": "输入格式: \"Bearer {token}\"",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}